refactor(example): cleanup android example after renames
