+++ packages/cloud_functions/cloud_functions/example/android/app/build.gradle
@@ -32,7 +32,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "io.flutter.plugins.firebase.cloudfunctions.cloudfunctionsexample"
+        applicationId "io.flutter.plugins.firebase.functions.example"
         minSdkVersion 16
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
