--- packages/cloud_functions/cloud_functions/example/functions/index.js
@@ -1,12 +1,0 @@
-const functions = require('firebase-functions');
-
-exports.listFruit = functions.https.onCall((data, context) => {
-   return [
-       "Apple",
-       "Banana",
-       "Cherry",
-       "Date",
-       "Fig",
-       "Grapes"
-   ]
-});
